\name{seleniumJars-package}
\alias{seleniumJars-package}
\alias{seleniumJars}
\docType{package}
\title{
JAR dependencies for the selenium package
}
\description{
This package provides external JAR dependencies for the selenium package. These jar files are needed for developing the selenium package using the rJava package, a low-level R to Java interface.
}
\author{
Lluis Ramon
}
